relocate froala svg
