+++ packages/modul-components/src/components/rich-text-editor/adapter/vue-froala.ts
@@ -8,12 +8,6 @@ import 'froala-editor/js/plugins.pkgd.min.js';
 import $ from 'jquery';
 import Component from 'vue-class-component';
 import { Emit, Prop, Watch } from 'vue-property-decorator';
-import boldIcon from '../../../assets/icons/svg/Froala-bold.svg';
-import listsIcon from '../../../assets/icons/svg/Froala-lists.svg';
-import replaceIcon from '../../../assets/icons/svg/Froala-replace.svg';
-import specialCharsIcon from '../../../assets/icons/svg/Froala-special-charracters.svg';
-import stylesIcon from '../../../assets/icons/svg/Froala-styles.svg';
-import titleIcon from '../../../assets/icons/svg/Froala-title.svg';
 import { ElementQueries } from '../../../mixins/element-queries/element-queries';
 import { replaceTags } from '../../../utils/clean/htmlClean';
 import { MFile } from '../../../utils/file/file';
@@ -22,6 +16,12 @@ import uuid from '../../../utils/uuid/uuid';
 import { ModulVue } from '../../../utils/vue/vue';
 import { FileUploadCustomValidation } from '../../file-upload/file-upload';
 import '../rte-internal.scss';
+import boldIcon from '../svg/Froala-bold.svg';
+import listsIcon from '../svg/Froala-lists.svg';
+import replaceIcon from '../svg/Froala-replace.svg';
+import specialCharsIcon from '../svg/Froala-special-charracters.svg';
+import stylesIcon from '../svg/Froala-styles.svg';
+import titleIcon from '../svg/Froala-title.svg';
 import { ImageLayoutCommands } from './image-layout-commands';
 import WithRender from './vue-froala.html?style=./vue-froala.scss';
